Rough cut of a happy path implementation
This commit contains initial work for a command-line tool that acts as a
replacement for `action-create-release-pr`.

Can you use this tool now? Yes! To do this:

* Clone this repo
* Check out the `create-tool` branch
* Run `yarn setup`
* Clone the `controllers` repo or find where you've already cloned it
* Check out the `monorepo-poc` branch
* Make sure your index is not dirty and you have nothing staged
* Run `node ../create-release-branch/dist/index.js`

This script should open your editor with a release spec template which
lists all packages in `controllers`. The template will allow you to
choose which packages in `controllers` you want to update. After you've
updated and closed this file, the script will bump the versions for the
packages you've specified and update their changelogs. It will then
determine a release version based on the current date and switch you to
a new branch based on this version.

There is a lot of remaining work to fully complete this tool, but here
is a rough summary:

* Instead of listing all packages in `controllers` (or whatever project
  you're running the tool against), list all packages for which commits
  exists since the last release which change the packages in any way.
* Sometimes the script throws an error. If it does, and packages have
  been bumped, then the script will fail when run again. The script
  should probably be idempotent.
* Add unit tests for existing code.
* Handle error cases which are marked via TODOs and anything else we can
  think of.
* Right now the tool is reusing some existing code in
  `action-create-release-pr` but in a weird way. We can integrate this
  code better.
* Ensure that polyrepos work just the same way they do now.