Reformat root version to include the release no.
This commit changes how the root version is bumped per release so that
instead of looking like `<yyyy>.<mm>.<dd>`, it looks like
`<yyyy><mm><dd>.<release-number>.0`. That is, the root version
consists of the date that the release was created + a release number,
which is incremented every time a new release is issued.