Update release spec to only list changed packages
Add a new property to the Package object,
`hasChangesSinceLatestRelease`. This property is computed when a package
is read from a project by running a diff between the last-created Git
tag which is associated with that package and whatever the HEAD commit
happens to be and checking to see whether any of the files changed
belong to that package; if so, then `hasChangesSinceLatestRelease` is
true, otherwise it's false.

There are a couple of things to consider here:

* "the last-created Git tag which is associated with that package" — how
  do we know this? How do we map a package's version to a tag? We have
  We have to account for tags that were created by
  `action-create-release-pr` in the past as well as the tags that this
  tool will create in the future. We also have to know what kind of
  package this is — whether it's the root package of a monorepo or a
  workspace package — because the set of possible tags will differ.
  These differences are documented in `readMonorepoRootPackage` and
  `readMonorepoWorkspacePackage`.
* What happens if a repo has no tags? Then all of the packages in that
  repo are considered to have changed and will thus get included within
  the release spec, since they need to receive their initial release.