Discourage postponing release of changed packages
When the tool generates a release spec, it will list all packages that
have changed since their last release. Editing the release spec to
remove a package so that it is no longer included in the release is
potentially dangerous, because it means that any package which relies on
that package that *is* included in the release could be broken in
production.

Since it's sometimes necessary to delay the release of a package, this
commit changes the release spec validation step such that if it detects
that a package that should be listed in the release spec *isn't*, it
will throw an error, advising the user of the danger, yet provide a
hidden option should the user really want to proceed.