Add functional tests (#19)
This commit adds tests which run the script within a separate shell
instance and verifies that the script makes the appropriate changes to
the repo.

I've tried to make these functional tests as readable as possible. In
order to do that, I created a small framework which replicates the
environment in which the script will run and then allows that
environment to be modified per-test to suit the assertions being made.