Reformat root version to include the release no. (#22)
This commit changes how the root version is represented so that instead
of looking like `<yyyy>.<mm>.<dd>`, it looks like
`<ordinary>.<backport>.0`. That is, for most releases, the first number
will get bumped, but for cases in which we want to issue a backport
release — a release which introduces a fix to a previous version — then
the second number will get incremented. The incrementing of the
"ordinary" number is only implemented in this PR; incrementing the
"backport" number will happen in a future PR.