Fix hasChangesInDirectorySinceGitTag (#39)
To determine whether a package needs a new release, we ask if there have
been any new commits that contain changes to any of the files within
that package. In other words, we run a `git diff` on the repo and look
for any files that match that package.

The problem is that `git diff` returns partial file paths, that is, file
paths which are relative to the repository directory, whereas the
subdirectory path that `hasChangesInDirectorySinceGitTag` takes is an
absolute path. So when we look for matching packages we always come up
short. This confuses the tool because it doesn't think there are any new
changes to release.