Fix published executable (#42)
Executables — files that ship with packages that are meant to be
executed and which are specified via the `bin` field in `package.json` —
aren't automatically run through Node. The standard way to do this is to
add a hashbang line to the executable.

Since we are using TypeScript, and we have to go through a compilation
step before we publish the package, we have two options:

* Add `#!/usr/bin/env node` to the top of `src/cli.ts`
* Make a new file *not* written in TypeScript which will execute the
  compiled version of the CLI

The former seems confusing, because it's impossible to run a TypeScript
file through `node`, so we go with the latter.