Update changelogs correctly for monorepo packages
Packages within a monorepo should now have their changelogs correctly
updated. Previously the update step was incorrectly expecting the
monorepo package to be using a tag of the format `v[version]`. This is
not how we tag packages within an indepentently versioned monorepo
because we need some way to distinguish the tags used for each project.

This step has been updated to specify the expected tag prefix as being
the package name followed by `@`.