Move auto-changelog to deps, pin to ~3.3.0
We started to use `@metamask/auto-changelog` at runtime in a recent
commit. As it turns out, recent versions of this tool contain breaking
changes that we didn't document, and so we need to rely on an older
version. Adding this to `dependencies` ensures that we are doing so,
even if the project where this tool is run against a project that also
also uses `@metamask/auto-changelog` (but a different version).