Move auto-changelog to deps, pin to ~3.3.0 (#122)
We started to use `@metamask/auto-changelog` at runtime in a recent
commit. As it turns out, recent versions of this tool contain breaking
changes that we didn't document, and so we need to rely on an older
version. Adding that version to `dependencies` ensures that we are
doing so, even if this tool is used in a project that also uses
`@metamask/auto-changelog` (but a different version).