Upgrade to MM `eslint-config-*` v15 + ESLint 9
The ESLint configuration is behind for this package. Keeping up to date
with our current lint rules and lint config format helps to debug issues
with ESLint now and in the future, and aligns this repo with our other
repos.

A list of changes:

- Upgrade `@metamask/eslint-config` and `@metamask/eslint-config-*`
  packages to 15.0.0
- Upgrade ESLint to 9.x
- Upgrade TypeScript ESLint packages to 8.25.x
- Upgrade other ESLint packages to fulfill peer dependencies
- Fix lint violations as a result of the upgrade

Note that the rules for `jsdoc/require-jsdoc` have been refined. In
performing this upgrade, many violations appeared that I thought didn't
help readability.