Fix repo links in updated changelogs to match `auto-changelog` (#165)
Fix repo links in updated changelogs

When this tool populates the Unreleased section of the changelog, it
does not just rewrite that section, but rather, it rewrites the entire
changelog. That behavior comes from the `updateChangelog` function in
`@metamask/auto-changelog`. This function takes the changes to apply,
but it also takes a repository URL, which it will use to construct links
for all of the releases in the changelog.

Unfortunately, the repository URL that this tool passes to
`updateChangelog` is not always the same URL that `auto-changelog` uses.
Specifically, if the developer used a different URL to clone the repo
than is listed in `package.json` under `repository.url`, then when
creating a new release branch, all of the links in the changelog will be
modified to reflect this new URL. This is incorrect, the resulting
changelog will not pass validation when run through `auto-changelog
validate`, forcing the developer to revert the changes to the links in
order to merge the release PR.

This commit addresses this pain point by updating the logic used to
obtain the repository URL to better match what is in
`@metamask/auto-changelog`. This means that the aforementioned links in
the changelog should stay the same when creating a new release branch.